backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
passlib[bcrypt]
bcrypt==4.0.1
python-jose==3.3.0
alembic==1.11.1
pydantic-settings==2.7.1
email-validator==2.1.0.post1
google-generativeai==0.3.2
python-multipart==0.0.20
httpx==0.25.0